+++ packages/icejs/bin/child-process-start.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
-const { childProcessStart } = require('create-cli-utils');
-const getBuiltInPlugins = require('../lib/index');
+const { getBuiltInPlugins, childProcessStart } = require('../lib');
 
 (async() => {
   await childProcessStart(getBuiltInPlugins);
