+++ packages/icejs/src/index.ts
@@ -1,42 +1,9 @@
-const getBuiltInPlugins = (userConfig) => {
-  if (userConfig.disableRuntime) {
-    return [
-      'build-plugin-react-app',
-      'build-plugin-ice-mpa'
-    ];
-  }
+/* eslint @typescript-eslint/no-var-requires: 0 */
+const { childProcessStart, createCli } = require('create-cli-utils');
+const getBuiltInPlugins = require('./getBuiltInPlugins');
 
-  const plugins = [
-    // common plugins
-    [
-      'build-plugin-app-core', {
-        'framework': 'react'
-      }
-    ],
-    'build-plugin-react-app',
-
-    // for ice/miniapp plugins
-    'build-plugin-miniapp',
-
-    // for ice/react plugins
-    'build-plugin-ice-router',
-    'build-plugin-ice-helpers',
-    'build-plugin-ice-logger',
-    'build-plugin-ice-config',
-    'build-plugin-ice-mpa',
-    'build-plugin-ice-request'
-  ];
-
-  if (userConfig.ssr) {
-    plugins.push('build-plugin-ice-ssr');
-  }
-
-  // add store plugin
-  if (!Object.prototype.hasOwnProperty.call(userConfig, 'store') || userConfig.store !== false) {
-    plugins.push('build-plugin-ice-store');
-  }
-
-  return plugins;
+module.exports = {
+  getBuiltInPlugins,
+  createCli,
+  childProcessStart
 };
-
-module.exports = getBuiltInPlugins;
